+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-9614 Added archivisation audit log
 * VFS-9902 Fixed GUI crash when a long path to file is shortened
 * VFS-9926 Redirecting from non-iframed Oneprovider GUI to provider view in Onezone GUI
 * VFS-9801 Added number of all processed (scanned) files in transfer